fix: Core project build success - resolved all 62 errors
- Complete canonical transformation of EncryptedConfiguration.cs
- Fixed CanonicalExecutorEnhanced validation and error handling
- Added missing methods to LatencyTracker (GetMetrics, GetPercentiles, Reset)
- Fixed PerformanceMonitor type conversions
- Resolved all TradingResult.ErrorMessage references
- Fixed CanonicalServiceBaseEnhanced method signatures
- Core project now builds with zero errors

This is a critical milestone as Core is a foundational dependency
for the entire solution. 364 errors remain in other projects.

+++ DayTradinPlatform/TradingPlatform.Core/Utilities/DecimalMath.cs
@@ -169,11 +169,11 @@ namespace TradingPlatform.Core.Utilities
             if (y == 0)
                 return 1;
             
-            if (x < 0 && y != Math.Floor((double)y))
+            if (x < 0 && y != (decimal)Math.Floor((double)y))
                 throw new ArgumentException("Negative base with non-integer exponent is not supported");
 
             // For integer exponents, use repeated multiplication
-            if (y == Math.Floor((double)y))
+            if (y == (decimal)Math.Floor((double)y))
             {
                 decimal result = 1;
                 decimal absY = Math.Abs(y);
